don't try and remove stdout!
git-svn-id: https://translate.svn.sourceforge.net/svnroot/translate/trunk/translate@404 54714841-351b-0410-a198-e36a94b762f5

+++ misc/optrecurse.py
@@ -352,8 +352,10 @@ class RecursiveOptionParser(optparse.OptionParser, object):
         self.finalizetempoutputfile(options, outputfile, fulloutputpath)
       return True
     else:
-      outputfile.close()
-      os.unlink(fulloutputpath)
+      # remove the file if it is a file (could be stdout etc)
+      if fulloutputpath and os.path.isfile(fulloutputpath):
+        outputfile.close()
+        os.unlink(fulloutputpath)
       return False
 
   def reportprogress(self, filename, success):
